created new file for markdown blog component
